Refactor EUR-Lex package: remove setup and requirements files, restructure codebase
- Deleted `requirements.txt` and `setup.py` as part of the refactor.
- Introduced new modules: `__init__.py`, `celex.py`, `constants.py`, `fetch.py`, `parser.py`, `sparql.py`, and `utils.py` to organize functionality.
- Implemented core functions for handling CELEX IDs, fetching HTML content, parsing articles, and executing SPARQL queries.
- Added utility functions for language normalization and IRI simplification.
- Created tests for various functionalities including SPARQL query execution, CELEX ID handling, and HTML parsing.
- Enhanced error handling and modularized code for better maintainability.
