+++ .gitignore
@@ -171,3 +171,13 @@ cython_debug/
 # Output data
 *.json
 *.html
+
+# Legacy packaging files
+requirements.txt
+setup.py
+MANIFEST.in
+.pypirc
+Pipfile
+
+# Generated by pytest-readme
+test_readme.py
