--- .pypirc
@@ -1,10 +1,0 @@
-[distutils]
-index-servers =
-    pypi
-    testpypi
-
-[pypi]
-repository = https://upload.pypi.org/legacy/
-
-[testpypi]
-repository = https://test.pypi.org/legacy/
