--- Pipfile
@@ -1,17 +1,0 @@
-[[source]]
-url = "https://pypi.org/simple"
-verify_ssl = true
-name = "pypi"
-
-[packages]
-sparqlwrapper = "==2.0.0"
-pandas = "==3.0.1"
-rdflib = "==7.6.0"
-requests = "==2.32.5"
-beautifulsoup4 = "==4.14.3"
-lxml = "==6.0.2"
-
-[dev-packages]
-
-[requires]
-python_version = "3.12"
