+++ README.md
@@ -3,6 +3,7 @@
 <p>
     <a href="https://github.com/kevin91nl/eurlex/actions/workflows/building.yaml"><img src="https://github.com/kevin91nl/eurlex/actions/workflows/building.yaml/badge.svg" alt="Building" height="18"></a>
     <a href="https://badge.fury.io/py/eurlex"><img src="https://badge.fury.io/py/eurlex.svg" alt="PyPI version" height="18"></a>
+    <a href="./LICENSE"><img src="https://img.shields.io/badge/license-MIT-blue.svg" alt="License: MIT" height="18"></a>
     <a href=https://github.com/ambv/black><img src="https://img.shields.io/badge/code%20style-black-000000.svg" height="18"></a>
 </p>
 
@@ -30,13 +31,11 @@ df = parse_html(html)
 df_article_1 = df[df.article == "1"]
 df_article_1_line_1 = df_article_1.iloc[0]
 
-# Display the subtitle of Article 1
-print(df_article_1_line_1.article_subtitle)
->>> "Subject matter"
-
-# Display the corresponding text
-print(df_article_1_line_1.text)
->>> "This Regulation lays down detailed provisions for the operation of unmanned aircraft systems as well as for personnel, including remote pilots and organisations involved in those operations."
+# Display the subtitle and corresponding text of Article 1
+assert df_article_1_line_1.article_subtitle == "Subject matter"
+assert df_article_1_line_1.text == (
+    "This Regulation lays down detailed provisions for the operation of unmanned aircraft systems as well as for personnel, including remote pilots and organisations involved in those operations."
+)
 ```
 
 Every document on EUR-Lex displays a CELEX number at the top of the page. More information on CELEX numbers can be found on the [EUR-Lex website](https://eur-lex.europa.eu/content/tools/eur-lex-celex-infographic-A3.pdf).
@@ -56,6 +55,26 @@ The following columns are available in the parsed dataframe:
 
 In some cases, additional fields are available. For example, the `group` field which contains the bold text under which a text is found.
 
+## Architecture
+
+The dependency graph below is generated by [`import-cruiser`](https://github.com/kevin91nl/import-cruiser) and refreshed by the pre-commit hook. It focuses on `src/eurlex` and its non-dev external dependencies, while keeping the current public import surface available through `eurlex`.
+
+![EUR-Lex dependency graph](./docs/architecture/dependency-graph.svg)
+
 ## Code Contribution
 
 Feel free to send any issues, ideas or pull requests.
+
+### Local checks
+
+For development, install the project and its hooks, then let pre-commit run the same checks that CI expects:
+
+```bash
+python -m pip install -e .[dev]
+pre-commit install
+pre-commit run --all-files
+```
+
+The final hook runs the doctests and enforces 100% coverage for `eurlex`, so you should see the same failures locally before a commit lands.
+
+The README examples are also exercised automatically through `pytest-readme`, so they stay in sync with the code instead of becoming decorative fiction.
