+++ conftest.py
@@ -1,0 +1,4 @@
+from pytest_readme import setup
+
+
+setup()
